Decouple the SDK from the buildpack (#63)
As discussed in https://salesforce.quip.com/F5g2ASm18WPG, this moves the `InvocationEvent`, `Context`, and `Logger` object creation into the SDK and out of the buildpack. There is a new function signature `enrichFn` that the buildpack can use to wrap the user's function with the enhanced functionality. 

This change is the complement to https://github.com/forcedotcom/nodejs-sf-fx-buildpack/pull/63

Co-authored-by: Travis <longoria.public@gmail.com>