Test types are importable
This PR adds a typescript file and then it exercises it to ensure that types are exported correctly via top level `npm run build`. In theory if something major is wrong with either the build process or the `src/index.ts` then this smoke test has a chance of catching it.