Test types are importable in CI (#82)
* Test types are importable

This PR adds a typescript file and then it exercises it to ensure that types are exported correctly via top level `npm run build`. In theory if something major is wrong with either the build process or the `src/index.ts` then this smoke test has a chance of catching it.

* Update test/fixtures/typescript-function/package.json

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

* Update .circleci/config.yml

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

* Resolve multiple node_modules in .gitignore

Node modules is listed several times in .gitignore. This declaration handles all our current cases.

* Test using typescript template

The prior test just checked that types import. This test includes a proper typescript salesforce function template. In addition it also executes the tests and linter against the project.

* Remove committed dist/ and adjust .gitignore to exclude it

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>